SchoolRecords/src/test/java/schoolRecords/TutorTest.java
package schoolRecords;

import org.junit.Test;
import schoolRecords.Subject;
import schoolRecords.Tutor;

import java.util.Arrays;

import static org.hamcrest.CoreMatchers.equalTo;
import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

public class TutorTest {

    @Test
    public void testTutorTeachingSubject() {
        //Given
        Tutor tutor = new Tutor("Nagy Csilla", Arrays.asList(new Subject("matematika"), new Subject("fizika")));
        //Then

        assertThat(tutor.getName(), equalTo("Nagy Csilla"));
        assertThat(tutor.tutorTeachingSubject(new Subject("fizika")), is(true));
    }

    @Test
    public void testTutorDoesNotTeachingSubject() {
        //Given
        Tutor tutor = new Tutor("Nagy Csilla", Arrays.asList(new Subject("matematika"), new Subject("fizika")));
        //Then

        assertThat(tutor.tutorTeachingSubject(new Subject("biológia")), is(false));
    }
}
